Switches s3 url to put bucket in path. S3 has a gotcha with their bucket URLs where if you have a bucket name with dots in them, it will be counted as a nested subdomain (of an already nested subdomain). For example, if you have a bucket called `s3-bucket`, you can access it at `https://s3-bucket.s3.us-east-1.amazonaws.com`, which works because AWS tls certs are for `*.s3.{region}.amazonaws.com`. But if your bucket name is something like `co.companyname.product`, the equivalent host is `https://co.companyname.product.s3.us-east-1.amazonaws.com`, which fails ssl validation due to a mismatched host in the cert (I guess wildcards don't support multiple subdomain levels deep?).
The alternative way to access the bucket is by putting the buket in the
url on the `s3.{region}.amazonaws.com` domain. So the above example
would be `https://s3.us-east-1.amazonaws.com/co.companyname.product`.
[Here's an
article](https://shlomoswidler.com/2009/08/amazon-s3-gotcha-using-virtual-host.html)
outlining the issue and the solution.

I also had to add "content-type" as a field in there because it seemed
that the `["starts-with", "$Content-Type", upload.mimetype],` was
failing because content-type wasn't being sent up.